Move {set,update}Scroll{Left,Top} to display/scrolling
