+++ src/edit/CodeMirror.js
@@ -4,7 +4,8 @@ import { setGuttersForLineNumbers, updateGutters } from "../display/gutters"
 import { maybeUpdateLineNumberWidth } from "../display/line_numbers"
 import { endOperation, operation, startOperation } from "../display/operations"
 import { initScrollbars } from "../display/scrollbars"
-import { onScrollWheel, setScrollLeft, updateScrollTop } from "../display/scroll_events"
+import { onScrollWheel } from "../display/scroll_events"
+import { setScrollLeft, updateScrollTop } from "../display/scrolling"
 import { clipPos, Pos } from "../line/pos"
 import { posFromMouse } from "../measurement/position_measurement"
 import { eventInWidget } from "../measurement/widgets"
